+++ etl/publish/publish_efficiency_metrics.py
@@ -3,7 +3,8 @@ import pandas as pd
 import numpy as np
 from database.utils import connect_to_db, insert_records, execute_query
 from utils import delete_published_records
-
+import os
+app_env = os.getenv("APP_ENV", "local")
 
 
 def read_records(tic):
@@ -59,7 +60,9 @@ def main():
     Main function to orchestrate the ETL process for efficiency metrics.
     """
     conn_local = connect_to_db("localhost")
-    conn_supabase = connect_to_db("supabase")
+    conn_supabase = True
+    if app_env == "local":
+        conn_supabase = connect_to_db("supabase")
     if conn_local and conn_supabase:
         cursor = conn_local.cursor()
         cursor.execute("SELECT tic FROM core.stock_profiles;")
@@ -67,7 +70,8 @@ def main():
         today = pd.Timestamp.now().date()
         try:
             total_deleted = delete_published_records(conn_local, "mart.efficiency_metrics", today, commit=False)
-            delete_published_records(conn_supabase, "mart.efficiency_metrics", today, commit=False)
+            if app_env == "local":
+                delete_published_records(conn_supabase, "mart.efficiency_metrics", today, commit=False)
             print(f"Deleted {total_deleted} records from mart.efficiency_metrics for as_of_date = {today}")
             for record in records:
                 tic = record[0]
@@ -86,21 +90,24 @@ def main():
                 total_inserted = insert_records(conn_local, df, "mart.efficiency_metrics", 
                                             ["tic", "date", "as_of_date"], 
                                             updated_at=False, commit=False)
-                insert_records(conn_supabase, df, "mart.efficiency_metrics", 
-                                            ["tic", "date", "as_of_date"], 
-                                            updated_at=False, commit=False)
+                if app_env == "local":
+                    insert_records(conn_supabase, df, "mart.efficiency_metrics", 
+                                                ["tic", "date", "as_of_date"], 
+                                                updated_at=False, commit=False)
                 print(f"For {tic}: Total records inserted = {total_inserted}")
         except Exception as e:
             print(f"Error processing efficiency metrics data: {e}")
             conn_local.rollback()
-            conn_supabase.rollback()
             conn_local.close()
-            conn_supabase.close()
+            if app_env == "local":
+                conn_supabase.rollback()
+                conn_supabase.close()
             return
         conn_local.commit()
         conn_local.close()
-        conn_supabase.commit()
-        conn_supabase.close()
+        if app_env == "local":
+            conn_supabase.commit()
+            conn_supabase.close()
         return
     return
