+++ etl/publish/publish_stock_scores.py
@@ -3,8 +3,9 @@ import pandas as pd
 import numpy as np
 from database.utils import connect_to_db, insert_records, execute_query
 from utils import delete_published_records
+import os 
 
-
+app_env = os.getenv("APP_ENV", "local")
 
 def read_records(tic):
     """
@@ -38,7 +39,10 @@ def main():
     Main function to orchestrate the ETL process for stock scores.
     """
     conn_local = connect_to_db("localhost")
-    conn_supabase = connect_to_db("supabase")
+    conn_supabase = True 
+    if app_env == "local":
+        conn_supabase = connect_to_db("supabase")
+         
     if conn_local and conn_supabase:
         cursor = conn_local.cursor()
         cursor.execute("SELECT tic FROM core.stock_profiles;")
@@ -46,7 +50,8 @@ def main():
         today = pd.Timestamp.now().date()
         try:
             total_deleted = delete_published_records(conn_local, "mart.stock_scores", today, commit=False)
-            delete_published_records(conn_supabase, "mart.stock_scores", today, commit=False)
+            if app_env == "local":
+                delete_published_records(conn_supabase, "mart.stock_scores", today, commit=False)
             print(f"Deleted {total_deleted} records from mart.stock_scores for as_of_date = {today}")
             for record in records:
                 tic = record[0]
@@ -59,21 +64,24 @@ def main():
                 total_inserted = insert_records(conn_local, df, "mart.stock_scores", 
                                             ["tic", "date", "as_of_date"], 
                                             updated_at=False, commit=False)
-                insert_records(conn_supabase, df, "mart.stock_scores", 
-                                            ["tic", "date", "as_of_date"], 
-                                            updated_at=False, commit=False)
+                if app_env == "local":
+                    insert_records(conn_supabase, df, "mart.stock_scores", 
+                                                ["tic", "date", "as_of_date"], 
+                                                updated_at=False, commit=False)
                 print(f"For {tic}: Total records inserted = {total_inserted}")
         except Exception as e:
             print(f"Error processing stock scores data: {e}")
             conn_local.rollback()
             conn_local.close()
-            conn_supabase.rollback()
-            conn_supabase.close()
+            if app_env == "local":
+                conn_supabase.rollback()
+                conn_supabase.close()
             return
         conn_local.commit()
         conn_local.close()
-        conn_supabase.commit()
-        conn_supabase.close()
+        if app_env == "local":
+            conn_supabase.commit()
+            conn_supabase.close()
         return
     return
